rec_sys/requirements.txt
psycopg2-binary
pandas==1.1.5
schedule
sqlalchemy==1.4.45